feat: auth module
로그인, 회원가입을 위해 auth module 을 생성 후 jwt 로그인, JwtAuthGuard, 커스텀 데코레이터 적용함
add Package
- bcrypt
- @nestjs/passport passport passport-local
- nestjs/jwt passport-jwt
- dev @types/passport-jwt

+++ server/src/modules/user/user.controller.ts
@@ -1,3 +1,5 @@
+import { CurrentUser } from './../../common/decorators/user.decorator';
+import { JwtAuthGuard } from './../auth/jwt/jwt.guard';
 import {
   Controller,
   Get,
@@ -6,14 +8,26 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { UserService } from './services/user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-
+import { LoginRequestDto } from '../auth/dto/login-request.dto';
+import { AuthService } from '../auth/auth.service';
+import { User } from './schemas/user.schema';
 @Controller('users')
 export class UserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly authService: AuthService,
+  ) {}
+
+  @Post('login')
+  logIn(@Body() data: LoginRequestDto): Promise<{ token: string }> {
+    return this.authService.jwtLogIn(data);
+  }
 
   @Post()
   create(@Body() createUserDto: CreateUserDto) {
@@ -30,16 +44,28 @@ export class UserController {
     return this.userService.findOneByUsername(username);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Patch(':username')
   update(
+    @CurrentUser() currentUser: User,
     @Param('username') username: string,
     @Body() updateUserDto: UpdateUserDto,
   ) {
+    if (currentUser.username !== username) {
+      throw new UnauthorizedException('접근 오류-본인만 접근할 수 있습니다');
+    }
     return this.userService.updateByUsername(username, updateUserDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':username')
-  remove(@Param('username') username: string) {
+  remove(
+    @CurrentUser() currentUser: User,
+    @Param('username') username: string,
+  ) {
+    if (currentUser.username !== username) {
+      throw new UnauthorizedException('접근 오류-본인만 접근할 수 있습니다');
+    }
     return this.userService.deleteOneByUsername(username);
   }
 }
